+++ CHANGELOG.md
@@ -16,3 +16,4 @@ The goal of this changelog is to track all changes made and to stand as the reco
 
 ## 0.0.3 - 2020/06/19
 1. Removed tooltip for the time being, was painful
+2. Fixed getUnusedAddress() development bug where crashed on live reload if account not loaded.